Rename: 폴더명 변경 (constant → constants)
