+++ src/components/ui/Spinner.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { ICON_SIZE } from '../../data/constant';
+import { ICON_SIZE } from '../../data/constants';
 
 import Spinner from './Spinner';
